+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
+### v0.2.0
 
-### v0.1
+* Add separated `parse` method.
+* Add request property `hasAuth`.
+
+### v0.1.0
 
 * Create basic authorization header parser.